requirements.txt
langchain_openai
langchain_core
langchain_community
langchain
streamlit
streamlit_float
langchain-pinecone 
pinecone-notebooks
PyPDF2
duckduckgo-search
python-dotenv
xformers
diffusers
transformers
accelerate
gradio
bitsandbytes
ftfy